refactor: ファイルを役割ごとに仕分ける
