+++ backend/app/src/main/kotlin/my/backend/Application.kt
@@ -1,0 +1,12 @@
+package my.backend
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+// --- Main Application ---
+@SpringBootApplication 
+class Application
+
+fun main(args: Array<String>) {
+    runApplication<Application>(*args)
+}
