+++ backend/app/src/main/kotlin/my/backend/config/WebConfig.kt
@@ -1,0 +1,22 @@
+package my.backend.config
+
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.context.annotation.Configuration
+import org.springframework.web.servlet.config.annotation.CorsRegistry
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
+
+// --- CORS Configuration ---
+@Configuration
+class WebConfig : WebMvcConfigurer {
+
+    @Value("\${cors.allowed-origins:http://localhost:3000}")
+    private lateinit var allowedOrigins: String
+
+    override fun addCorsMappings(registry: CorsRegistry) {
+        registry.addMapping("/api/**")
+                .allowedOrigins(allowedOrigins)
+                .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS")
+                .allowedHeaders("*")
+                .allowCredentials(true)
+    }
+}
